Implemented Add, Edit, Delete, View all vehicles ------------------------------------------------ Implementation of Add vehicles, including ability to add an image.
Implementation of Edit a selected vehicle, with validation, the
image for the vehicle is only overwritten if there is a new image
selected for uploading.

Deletion of a vehicle forwards the admin back to the list of
vehicles.
Deletion of a vehicle still requires the addition of if the vehicle
is booked or not, to be implemented when booking model is created.

View a list of all vehicles, shows rows of vehicles, displaying an
image of the vehicle if it's given, or if not, depending on the
body type of the vehicle, a generic image is shown.